Changes to set config properties during download (#16)
* Changes to set config properties during download

* Include phi 1.5 in models
