+++ README.md
@@ -79,13 +79,12 @@ python3 $WORK_DIR/llm/download.py --model_name llama2_7b --output /mnt/llm --hf_
 
 Run the following command for starting Kubeflow serving and running inference on the given input:
 ```
-bash run.sh  -n <MODEL_NAME> -d <INPUT_PATH> -g <NUM_GPUS> -m <NFS_LOCAL_MOUNT_LOCATION> -f <NFS_ADDRESS_WITH_SHARE_PATH> -e <KUBE_DEPLOYMENT_NAME> [OPTIONAL -k]
+bash run.sh  -n <MODEL_NAME> -d <INPUT_PATH> -g <NUM_GPUS> -f <NFS_ADDRESS_WITH_SHARE_PATH> -e <KUBE_DEPLOYMENT_NAME> [OPTIONAL -k]
 ```
 - k:    Set flag to keep server alive
 - n:    Name of model
 - d:    Absolute path of input data folder
 - g:    Number of gpus to be used to execute (Set 0 to use cpu)
-- m:    Absolute path to the NFS local mount location
 - f:    NFS server address with share path information
 - e:    Name of the deployment metadata
 
@@ -97,15 +96,15 @@ Should print "Inference Run Successful" as a message at the end
 
 For 1 GPU Inference with official MPT-7B model and keep inference server alive:
 ```
-bash $WORK_DIR/llm/run.sh -n mpt_7b -d data/translate -m /mnt/llm -g 1 -e llm-deploy -f '1.1.1.1:/llm' -k
+bash $WORK_DIR/llm/run.sh -n mpt_7b -d data/translate -g 1 -e llm-deploy -f '1.1.1.1:/llm' -k
 ```
 For 1 GPU Inference with official Falcon-7B model and keep inference server alive:
 ```
-bash $WORK_DIR/llm/run.sh -n falcon_7b -d data/qa -m /mnt/llm -g 1 -e llm-deploy -f '1.1.1.1:/llm' -k
+bash $WORK_DIR/llm/run.sh -n falcon_7b -d data/qa -g 1 -e llm-deploy -f '1.1.1.1:/llm' -k
 ```
 For 1 GPU Inference with official Llama2-7B model and keep inference server alive:
 ```
-bash $WORK_DIR/llm/run.sh -n llama2_7b -d data/summarize -m /mnt/llm -g 1 -e llm-deploy -f '1.1.1.1:/llm' -k
+bash $WORK_DIR/llm/run.sh -n llama2_7b -d data/summarize -g 1 -e llm-deploy -f '1.1.1.1:/llm' -k
 ```
 
 #### Inference Check
