+++ llm/model_config.json
@@ -55,5 +55,18 @@
             "response_timeout" : 2000
         },
         "gpu_type": ["A100-PCIE-40GB","A100-PCIE-80GB"]
+    },
+    "phi-1_5": {
+        "model_name": "phi-1_5",
+        "version": "1.0",
+        "handler": "handler.py",
+        "repo_id": "microsoft/phi-1_5",
+        "requirements_file": "model_requirements.txt",
+        "registration_params":{
+            "batch_size" : 16,
+            "max_batch_delay" : 2000,
+            "response_timeout" : 2000
+        },
+        "gpu_type": ["A100-PCIE-40GB","A100-PCIE-80GB"]
     }
 }
